+++ src/Units/DoRUnits.java
@@ -149,7 +149,7 @@ public class DoRUnits extends UnitModelScheme
     }
 
     @Override
-    public double getDamageRedirect(WeaponModel wm)
+    public int getDamageRedirect(WeaponModel wm)
     {
       return wm.getDamage(this);
     }
